requirements.txt
 streamlit
  anthropic
  sentence-transformers
  numpy
  python-dotenv
  torch
  torchvision